#6-add-views 파일 하나로 통합, urls.py 수정
